DONE: - Удалил класс DBUserDetails, заиплементил интерфейс UserDetails к Customer; - Реализованы методы интерфейса, к ролям добавил FetchType.EAGER для загрузки в БД ролей сразу с пользователями, чтоб не ждать когда к ним обратятся; - Role заимплементил GrantedAuthority, для корректной передачи в Customer, чтоб потом можно было вытянуть из колекции; - ViewController добавлены комиты и переписан метод проверки на Админ. TODO: - Добавить все связи с сущностями(Event, Activity). - Добавить к сущностям(Event, Activity) Dao, Repo, Service - Добавить в BD - UT5 желательно бы :)
