--- src/main/java/org/stepacademy/swm_diplom_mvc/model/entities/customer/dbUserDetails/DBUserDetails.java
@@ -1,55 +1,0 @@
-package org.stepacademy.swm_diplom_mvc.model.entities.customer.dbUserDetails;
-
-import lombok.Data;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.stepacademy.swm_diplom_mvc.model.entities.customer.customer.Customer;
-
-import java.util.Collection;
-import java.util.Collections;
-
-@Data
-public class DBUserDetails implements UserDetails {
-    private Customer customer;
-
-
-    public DBUserDetails(Customer customer) {
-        this.customer = customer;
-    }
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return Collections.<GrantedAuthority>singletonList(new SimpleGrantedAuthority("ROLE_ADMIN"));
-    }
-
-    @Override
-    public String getPassword() {
-        return customer.getPassword();
-    }
-
-    @Override
-    public String getUsername() {
-        return customer.getLogin();
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-}
