+++ src/main/java/org/stepacademy/swm_diplom_mvc/model/entities/customer/role/Role.java
@@ -3,6 +3,7 @@ package org.stepacademy.swm_diplom_mvc.model.entities.customer.role;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
+import org.springframework.security.core.GrantedAuthority;
 import org.stepacademy.swm_diplom_mvc.model.entities.customer.customer.Customer;
 
 import javax.management.ConstructorParameters;
@@ -13,7 +14,7 @@ import java.util.Set;
 @Setter
 @Entity
 @Table(name = "role_t")
-public class Role {
+public class Role implements GrantedAuthority {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
@@ -31,5 +32,17 @@ public class Role {
     public Role(Integer id, String role) {
         this.id = id;
         this.role = role;
+    }
+
+    @Override
+    public String getAuthority() {
+        return getRole();
+    }
+
+    @Override
+    public String toString() {
+        return "Role{" +
+                "role='" + role + '\'' +
+                '}';
     }
 }
