+++ src/main/resources/templates/layouts/header.html
@@ -22,6 +22,7 @@
     <button class="button button--0">new event</button>
     <button class="button button--1">chat</button>
 
+<!--    <th:block th:each=""-->
     <button class="button" th:if="${session.isAdmin}">
         admin page
     </button>
